app/celery_worker.py
from celery import Celery
import os
from dotenv import load_dotenv

load_dotenv()

# Redis URL (লোকাল টেস্টিংয়ের জন্য ডিফল্ট পোর্ট 6379)
REDIS_URL = os.getenv("REDIS_URL", "redis://localhost:6379/0")

celery_app = Celery(
    "prohory_scanner",
    broker=REDIS_URL,
    backend=REDIS_URL
)

celery_app.conf.update(
    task_serializer="json",
    result_serializer="json",
    accept_content=["json"],
)

@celery_app.task
def background_scan_task(message: str, webhook_url: str):
    """
    এই টাস্কটি ব্যাকগ্রাউন্ডে চলবে। 
    এখানেই আমরা scanner.py এর লজিক কল করব।
    """
    print(f"Starting deep scan for message: {message}")
    
    # TODO: এখানে GSB, VirusTotal, HF Model এবং Scraping লজিক বসবে
    # scan_result = process_message_hybrid(message)
    
    # ডামি রেজাল্ট
    scan_result = {"status": "SAFE", "confidence": 0.85}
    
    # TODO: স্ক্যান শেষে Webhook_url এ requests.post() করে Spring Boot কে রেজাল্ট জানানো হবে
    print(f"Scan complete. Sending result to {webhook_url}")
    return scan_result